+++ include/minigui.h
@@ -1958,7 +1958,7 @@ const CompositorOps* __ex_compositor_get (const char* name,
  *
  * \sa RegisterRequestHandler
  */
-#define MAX_SYS_REQID           0x0023
+#define MAX_SYS_REQID           0x0024
 
 /**
  * \def MAX_REQID